Move Spinner & Loader to flatten directory structure
