+++ packages/client/src/components/form/DocumentUploadfield/DocumentListPreview.tsx
@@ -13,7 +13,7 @@ import * as React from 'react'
 import styled from 'styled-components'
 import { PaperClip, Delete } from '@opencrvs/components/lib/icons'
 import { IFileValue, IAttachmentValue, IFormFieldValue } from '@client/forms'
-import { Spinner } from '@opencrvs/components/lib/interface'
+import { Spinner } from '@opencrvs/components/lib/Spinner'
 import { withTheme, ITheme } from '@client/styledComponents'
 import { ISelectOption } from '@opencrvs/components/lib/Select'
 import { Button } from '@opencrvs/components/lib/buttons/Button'
