+++ packages/client/src/components/form/FetchButton.tsx
@@ -22,7 +22,7 @@ import {
   IntlShape
 } from 'react-intl'
 import { buttonMessages } from '@client/i18n/messages'
-import { Spinner } from '@opencrvs/components/lib/interface'
+import { Spinner } from '@opencrvs/components/lib/Spinner'
 import { PrimaryButton } from '@opencrvs/components/lib/buttons'
 import { Success, Error } from '@opencrvs/components/lib/icons'
 import { IQuery } from '@opencrvs/client/src/forms'
