+++ packages/client/src/views/EventInfo/EventInfo.tsx
@@ -28,7 +28,7 @@ import {
   PrimaryButton
 } from '@opencrvs/components/lib/buttons'
 import { BackArrow } from '@opencrvs/components/lib/icons'
-import { BulletList } from '@opencrvs/components/lib/typography'
+import { BulletList } from '@opencrvs/components/lib/BulletList'
 import { connect } from 'react-redux'
 import {
   goBack,
