+++ packages/client/src/views/OfficeHome/OfficeHome.tsx
@@ -44,9 +44,9 @@ import {
 } from '@client/utils/constants'
 import {
   FloatingNotification,
-  NOTIFICATION_TYPE,
-  Spinner
+  NOTIFICATION_TYPE
 } from '@opencrvs/components/lib/interface'
+import { Spinner } from '@opencrvs/components/lib/Spinner'
 import * as React from 'react'
 import { injectIntl, WrappedComponentProps as IntlShapeProps } from 'react-intl'
 import { connect } from 'react-redux'
