+++ packages/client/src/views/OfficeHome/Outbox.tsx
@@ -11,7 +11,7 @@
  */
 import * as React from 'react'
 import { BodyContent } from '@opencrvs/components/lib/Content'
-import { Spinner } from '@opencrvs/components/lib/interface'
+import { Spinner } from '@opencrvs/components/lib/Spinner'
 import {
   Workqueue,
   ColumnContentAlignment
