+++ packages/client/src/views/OfficeHome/components.tsx
@@ -15,7 +15,7 @@ import { DeclarationIcon } from '@opencrvs/components/lib/icons/DeclarationIcon'
 import { STATUSTOCOLOR } from '@client/views/RecordAudit/RecordAudit'
 import { Duplicate, StatusFailed } from '@opencrvs/components/lib/icons'
 import { SUBMISSION_STATUS } from '@client/declarations'
-import { Spinner } from '@opencrvs/components/lib/interface/Spinner'
+import { Spinner } from '@opencrvs/components/lib/Spinner'
 import { Uploaded } from '@opencrvs/components/lib/icons/Uploaded'
 import { WaitingToSent } from '@opencrvs/components/lib/icons/WaitingToSent'
 import { ConnectionError } from '@opencrvs/components/lib/icons/ConnectionError'
