+++ packages/client/src/views/SysAdmin/Team/user/userCreation/CreateNewUser.tsx
@@ -30,7 +30,8 @@ import {
 import { replaceInitialValues } from '@client/views/RegisterForm/RegisterForm'
 import { UserForm } from '@client/views/SysAdmin/Team/user/userCreation/UserForm'
 import { UserReviewForm } from '@client/views/SysAdmin/Team/user/userCreation/UserReviewForm'
-import { ActionPageLight, Spinner } from '@opencrvs/components/lib/interface'
+import { ActionPageLight } from '@opencrvs/components/lib/interface'
+import { Spinner } from '@opencrvs/components/lib/Spinner'
 import ApolloClient from 'apollo-client'
 import * as React from 'react'
 import { withApollo } from 'react-apollo'
