+++ packages/components/src/interface/ExpandableNotification/ExpandableNotification.tsx
@@ -12,7 +12,7 @@
 import * as React from 'react'
 import styled, { withTheme } from 'styled-components'
 import { Button } from '../../buttons'
-import { Spinner } from '../Spinner'
+import { Spinner } from '../../Spinner'
 import { KeyboardArrowUp, KeyboardArrowDown, Outbox } from '../../icons'
 import { colors } from '../../colors'
 import { ITheme } from '../../theme'
