+++ packages/components/src/interface/index.ts
@@ -11,7 +11,6 @@
  */
 export * from './Box'
 export * from './Header'
-export * from './Spinner'
 export * from './Tabs'
 export * from './Modal'
 export * from './FloatingNotification'
@@ -27,7 +26,6 @@ export * from './PINKeypad'
 export * from './SearchTool'
 export * from './ExpandingMenu/ExpandingMenu'
 export * from './ExpandableNotification/ExpandableNotification'
-export * from './Loader'
 export * from './Header/AppHeader'
 export * from './Header/PageHeader'
 export * from './ToggleMenu'
